requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
httpx>=0.24.0
pyyaml>=6.0
pydantic>=2.0
python-multipart
